pyproject.toml
[build-system]
requires = ["setuptools>=67"]
build-backend = "setuptools.build_meta"

[project]
name = "notionapi"
dynamic = ["version"]
authors = [
  { name="Matteo Beatrice", email="matteo.beatrice@sentric.it" },
]
description = "Python package used to automate operations with Notion"
readme = {file = "README.txt", content-type = "text/markdown"}
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "PyGithub>=1.58.2, <2",
  "notion-client>=2.0.0, <3",
  "toml>=0.10.2, <1"
]

[tool.setuptools.dynamic]
version = {attr = "notionapi.VERSION"}

[project.urls]
"Homepage" = "https://github.com/sentricsrl/notion-api"
"Bug Tracker" = "https://github.com/sentricsrl/notion-api/issues"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
